+++ compiler/types.cc
@@ -212,5 +212,14 @@ std::string TargetPlatformSLExtension(TargetPlatform platform) {
   FML_UNREACHABLE();
 }
 
+std::string ToUtf8(const std::wstring& wstring) {
+  std::wstring_convert<std::codecvt_utf8<wchar_t>> myconv;
+  return myconv.to_bytes(wstring);
+}
+
+std::string ToUtf8(const std::string& string) {
+  return string;
+}
+
 }  // namespace compiler
 }  // namespace impeller
